feat(landing): promote v3 to / canonical, preserve legacy at /v2
Same migration pattern used for the v2 app surfaces — flip the new design
to the canonical path, preserve the old one at a versioned alias for
rollback or A/B.

Rename:
  src/features/landing/    → src/features/landing-v2/    (legacy editorial)
  src/features/landing-v3/ → src/features/landing/       (new editorial)
  landing/LandingV3.jsx    → landing/Landing.jsx         (component renamed
                                                          to Landing, CSS
                                                          import updated to
                                                          ./landing.css)

Routes:
  /     → Landing  (was the v2 design, now the v3 design)
  /v2   → LandingV2 (legacy editorial, preserved)
  /v3   → removed (the route the prototype rolled out under is now gone)

Cross-folder consumers updated to point at the relocated legacy code:
  src/app/AppShell.jsx                — useGoogleAuth
  src/app/header/Header.jsx           — useGoogleAuth
  src/app/pages/legal/AboutPage.jsx   — useGoogleAuth + TopNav + Footer
  src/app/pages/legal/PrivacyPage.jsx — TopNav + Footer
  src/app/pages/legal/TermsPage.jsx   — TopNav + Footer
  src/app/router.jsx                  — Landing/LandingV2 lazy imports +
                                        legal-page TopNav + Footer

Inside landing-v2/ every `@/features/landing/` self-reference rewritten
to `@/features/landing-v2/` so the folder is self-contained (the new
landing/ uses its own utils-free inline structure plus the legacy
useGoogleAuth via landing-v2).

The v3 landing's FEELFLICK wordmark anchor now points at / (was /v3).

Verification: lint clean, 519 tests pass, prod build OK, dev smoke:
  /   → v3 editorial spread (11,550px, "Films that know you.")
  /v2 → legacy editorial ("The right film. Right now.")
  /v3 → 404 page
