+++ src/app/header/Header.jsx
@@ -13,7 +13,7 @@ import {
 
 import { supabase } from '@/shared/lib/supabase/client'
 import { useAuthSession } from '@/shared/hooks/useAuthSession'
-import { useGoogleAuth } from '@/features/landing/utils/useGoogleAuth'
+import { useGoogleAuth } from '@/features/landing-v2/utils/useGoogleAuth'
 import { useUnreadFeed } from '@/shared/hooks/useUnreadFeed'
 
 // Mood-tinted accent. Wire to a context later (see notes in README).
