+++ src/features/landing-v2/components/Footer.jsx
