+++ src/features/landing-v2/sections/MoatProofSection.jsx
