+++ src/features/landing-v2/sections/MoodShowcaseSection.jsx
